// convex/_generated/api.d.ts
/* eslint-disable */
/**
 * Generated `api` utility.
 *
 * THIS CODE IS AUTOMATICALLY GENERATED.
 *
 * To regenerate, run `npx convex dev`.
 * @module
 */

import type {
  ApiFromModules,
  FilterApi,
  FunctionReference,
} from "convex/server";
import type * as collab from "../collab.js";
import type * as github from "../github.js";
import type * as http from "../http.js";
import type * as myFunctions from "../myFunctions.js";
import type * as snippet from "../snippet.js";
import type * as users from "../users.js";
import type * as videoCall from "../videoCall.js";

/**
 * A utility for referencing Convex functions in your app's API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
declare const fullApi: ApiFromModules<{
  collab: typeof collab;
  github: typeof github;
  http: typeof http;
  myFunctions: typeof myFunctions;
  snippet: typeof snippet;
  users: typeof users;
  videoCall: typeof videoCall;
}>;
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;